requirements.txt
opencv-python
ultralytics
mediapipe==0.10.11
pynput 
numpy
pycaw 
comtypes
customtkinter 
pillow
python-dotenv
screen-brightness-control
#pip install --extra-index-url https://PySimpleGUI.net/install PySimpleGUI